much faster TreMCFilterCriterion
git-svn-id: http://svn.cern.ch/guest/lhcb/Analysis/trunk@70392 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 9c770ec187cfa9bb3feb3fe627e4414195af5f16

+++ Analysis/Phys/DaVinciMCTools/src/TrueMCFilterCriterion.cpp
@@ -1,4 +1,4 @@
-// $Id: TrueMCFilterCriterion.cpp,v 1.18 2007-07-05 13:09:58 pkoppenb Exp $
+// $Id: TrueMCFilterCriterion.cpp,v 1.19 2007-07-24 11:47:14 pkoppenb Exp $
 // Include files 
 
 // from Gaudi
@@ -6,12 +6,6 @@
 
 #include "MCInterfaces/IMCDecayFinder.h"
 #include "Kernel/Particle2MCLinker.h"
-
-#include "Kernel/CaloCellID.h"
-#include "Event/CaloCluster.h"
-#include "CaloKernel/CaloVector.h"
-#include "CaloUtils/Calo2MC.h"
-//#include "Event/TrgCaloParticle.h"
 
 // local
 #include "TrueMCFilterCriterion.h"
@@ -21,7 +15,6 @@
 //
 // 2004-09-13 : Patrick KOPPENBURG
 // 
-// Reviewed to have gammas 2005-03-20 : Luis Fernandez
 //-----------------------------------------------------------------------------
 
 // Declaration of the Tool Factory
@@ -37,14 +30,14 @@ TrueMCFilterCriterion::TrueMCFilterCriterion( const std::string& type,
   : FilterCriterionBase ( type, name , parent )
   , m_pMCDecFinder(0)
   , m_pLinker(0)
-  , m_filterOut(false)
-  , m_pCl2MCPTable(0)
+  , m_mcParticles(0)
+  , m_eventCount(0)
 {
   declareInterface<IFilterCriterion>(this);
-  declareProperty( "VetoSignal", m_filterOut );
-  declareProperty( "ParticlePath", 
-                   m_particlePaths );
-  
+  declareProperty( "VetoSignal", m_filterOut = false );
+  declareProperty( "ParticlePath", m_particlePaths );
+  declareProperty( "MCParticlePath", m_mcParticlePath = LHCb::MCParticleLocation::Default);
+  m_decayMembers.clear();
 }
 //=============================================================================
 // Destructor
@@ -71,11 +64,6 @@ StatusCode TrueMCFilterCriterion::initialize( ){
                                     Particle2MCMethod::Chi2,
                                     m_particlePaths);
 
-  // for calo clusters -> MCPatricle association
-  // Not actually used anywhere
-//   m_pCl2MCPTable = get<Table>("Relations/" + 
-//                               LHCb::CaloClusterLocation::Default);
-
   if (msgLevel(MSG::VERBOSE)) verbose() << "Initialised happily" << endmsg ;
   setActive(); 
 
@@ -88,35 +76,79 @@ bool TrueMCFilterCriterion::testParticle( const LHCb::Particle* const & part ) {
  
   if (msgLevel(MSG::VERBOSE)) verbose() << "TrueMCFilterCriterion" << endmsg ;
 
-  const LHCb::MCParticle* MC = m_pLinker->firstMCP( part );
-  
-  if( NULL == MC ){
-    if (msgLevel(MSG::DEBUG)) debug() << "Empty association range" << endmsg;
-    if (msgLevel(MSG::DEBUG)) debug() << "No association for " << part->key() << " "
-                                      << part->particleID().pid() << " " << part->momentum()  << endmsg ;
-    return m_filterOut ; // true if one wants to kill all, false else
-  }
-
-  if ( msgLevel ( MSG::DEBUG )) debug() << "Particle      " << part->key() << " " 
-                                        << part->particleID().pid() << " " 
-                                        << part->momentum() 
-                                        << endmsg ;
-
   bool signal = false;
 
-  while ( NULL != MC ){
+  if (!preloadMCParticles()){ // preload particles
+    if (msgLevel(MSG::DEBUG)) debug() << "Signal not found" << endmsg ;
+    signal = false ;
+  } else {    
 
-    if (msgLevel(MSG::DEBUG)) debug() << "Associated to " << MC->key() << " " 
-                                      << MC->particleID().pid() << " " 
-                                      << MC->momentum() << endmsg ;
-    signal = findMCParticle(MC);
+    const LHCb::MCParticle* MC = m_pLinker->firstMCP( part );
+  
+    if( NULL == MC ){
+      if (msgLevel(MSG::DEBUG)) debug() << "Empty association range" << endmsg;
+      if (msgLevel(MSG::DEBUG)) debug() << "No association for " << part->key() << " "
+                                        << part->particleID().pid() << " " << part->momentum()  << endmsg ;
+      return m_filterOut ; // true if one wants to kill all, false else
+    }
     
-    if(signal) if (msgLevel(MSG::DEBUG)) debug() << "which is a signal particle" << endmsg;
-    if(signal) break; // just take one associated MCParticle
-    MC = m_pLinker->nextMCP();
+    if ( msgLevel ( MSG::DEBUG )) debug() << "Particle      " << part->key() << " " 
+                                          << part->particleID().pid() << " " 
+                                          << part->momentum() 
+                                          << endmsg ;
+    
+    while ( NULL != MC ){
+      
+      if (msgLevel(MSG::DEBUG)) debug() << "Associated to " << MC->key() << " " 
+                                        << MC->particleID().pid() << " " 
+                                        << MC->momentum() << endmsg ;
+      signal = findMCParticle(MC);
+      
+      if(signal) if (msgLevel(MSG::DEBUG)) debug() << "which is a signal particle" << endmsg;
+      if(signal) break; // just take one associated MCParticle
+      MC = m_pLinker->nextMCP();
+    }
   }
-
+  
   return ( signal ) ? (!m_filterOut) : m_filterOut ;
+  
+}
+
+//=========================================================================
+// Call DecayFinder only once per event
+//=========================================================================
+bool TrueMCFilterCriterion::preloadMCParticles ( ) {
+  // MC list
+  const LHCb::MCParticles* kmcparts = get<LHCb::MCParticles>( m_mcParticlePath );
+  if( !kmcparts ){
+    fatal() << "Unable to find MC particles at '" << m_mcParticlePath << "'" << endmsg;
+    return false;
+  }
+  if ( kmcparts != m_mcParticles ){ // new event -> load particles
+    if (msgLevel(MSG::DEBUG)) debug() << "New event " << endmsg ;
+    m_eventCount++;
+    m_mcParticles = kmcparts ;
+    m_decayMembers.clear() ; // clear decay members
+    
+    LHCb::MCParticle::ConstVector MCHead;
+    const LHCb::MCParticle* imc = NULL;
+    while ( m_pMCDecFinder->findDecay(*kmcparts, imc ) ){
+      if (msgLevel(MSG::DEBUG)) debug() << "Found decay head " << imc->key() << " " 
+                                        << imc->particleID().pid() << endmsg ;
+      MCHead.push_back(imc);
+    }
+    LHCb::MCParticle::ConstVector::const_iterator ihead;
+    for( ihead = MCHead.begin(); ihead != MCHead.end(); ++ihead){
+      const LHCb::MCParticle* mc = *ihead;
+      // final states must be flagged
+      m_pMCDecFinder->decayMembers(mc, m_decayMembers); 
+    }
+    if (msgLevel(MSG::DEBUG)) debug() << "Found " << m_decayMembers.size() 
+                                      << " MC particles from true decay" << endmsg ;
+  } else if (msgLevel(MSG::DEBUG)) debug() << "Reusing " << m_decayMembers.size() 
+                                           << " MC particles from true decay" << endmsg ;
+  
+  return !(m_decayMembers.empty()) ;  
   
 }
 //=============================================================================
@@ -124,38 +156,13 @@ bool TrueMCFilterCriterion::testParticle( const LHCb::Particle* const & part ) {
 //=============================================================================
 bool TrueMCFilterCriterion::findMCParticle( const LHCb::MCParticle* MC ) {
 
-  if (msgLevel(MSG::VERBOSE)) verbose() << "TrueMCFilterCriterion find MC" << endmsg ;
-
-  // MC list
-  LHCb::MCParticle::Vector mclist; // LF
-  LHCb::MCParticles* kmcparts = get<LHCb::MCParticles>(LHCb::MCParticleLocation::Default );
-  if( !kmcparts ){
-    fatal() << "Unable to find MC particles at '" << LHCb::MCParticleLocation::Default << "'" << endmsg;
-    return false;
-  }
-
-  LHCb::MCParticle::ConstVector MCHead;
-  const LHCb::MCParticle* imc = NULL;
-  while ( m_pMCDecFinder->findDecay(*kmcparts, imc ) ){
-    const LHCb::MCParticle* jmc = imc;
-    MCHead.push_back(jmc);
-  }
-   
-  LHCb::MCParticle::ConstVector::const_iterator ihead;
-  for( ihead = MCHead.begin(); ihead != MCHead.end(); ++ihead){
-    const LHCb::MCParticle* mc = *ihead;
-    // final states must be flagged
-    m_pMCDecFinder->decayMembers(mc, mclist);
-     
-  }
-  
-  if (msgLevel(MSG::VERBOSE)) verbose() << "Found " << mclist.size() << " MC particles from true decay" << endmsg ;
-  if (mclist.empty()) return false ;
+  if (msgLevel(MSG::VERBOSE)) verbose() << "TrueMCFilterCriterion find MC" << endmsg ;  
+  if (m_decayMembers.empty()) return false ;
 
   if (msgLevel(MSG::VERBOSE)) verbose() << "Looking for " << MC->key() 
                                         << " " << MC->particleID().pid() << " " << MC->momentum() << endmsg ;  
   LHCb::MCParticle::Vector::iterator mcp ; // LF
-  for ( mcp = mclist.begin() ; mcp != mclist.end() ; ++mcp ){
+  for ( mcp = m_decayMembers.begin() ; mcp != m_decayMembers.end() ; ++mcp ){
     if (msgLevel(MSG::VERBOSE)) verbose() << "Looping on  " << (*mcp)->key() 
                                           << " " << (*mcp)->particleID().pid() << " " 
                                           << (*mcp)->momentum() << endmsg ;  
@@ -172,6 +179,7 @@ bool TrueMCFilterCriterion::findMCParticle( const LHCb::MCParticle* MC ) {
 StatusCode TrueMCFilterCriterion::finalize(){
   
   if( NULL != m_pLinker ) delete m_pLinker;
+  always() << "Tested " << m_eventCount << " events" << endmsg ;
   return GaudiTool::finalize() ;
   
 }
